+++ app/helpers/application_helper.rb
@@ -65,4 +65,15 @@ module ApplicationHelper
     return conversions_voluntary_project_assign_caseworker_path(project) if project.task_list_type == "Conversion::Voluntary::TaskList"
     return conversions_involuntary_project_assign_caseworker_path(project) if project.task_list_type == "Conversion::Involuntary::TaskList"
   end
+
+  def optional_cookies_set?
+    cookies[:ACCEPT_OPTIONAL_COOKIES].present?
+  end
+
+  def enable_google_tag_manager?
+    return false unless ENV["SENTRY_ENV"] == "production"
+    return false unless ENV["GOOGLE_TAG_MANAGER_ID"].present?
+    return false unless cookies[:ACCEPT_OPTIONAL_COOKIES] == true
+    true
+  end
 end