+++ config/application.rb
@@ -39,5 +39,8 @@ module DfeCompleteConversionsTransfersAndChanges
     config.exceptions_app = routes
 
     config.support_email = "regionalservices.rg@education.gov.uk"
+
+    # use the cookie session and set the name of the cookie
+    config.session_store :cookie_store, key: "SESSION"
   end
 end