+++ config/routes.rb
@@ -92,6 +92,9 @@ Rails.application.routes.draw do
     end
   end
 
+  get "cookies", to: "cookies#edit"
+  post "cookies", to: "cookies#update"
+
   # High voltage configuration for static pages. Matches routes from the root of the domain. Uses
   # HighVoltage::Constraints::RootRoute to validate that the view exists.
   get "/*id" => "pages#show", :as => :page, :format => false, :constraints => HighVoltage::Constraints::RootRoute.new