+++ spec/rails_helper.rb
@@ -79,7 +79,6 @@ RSpec.configure do |config|
   config.include AcademiesApiHelpers
   config.include FeatureHelpers, type: :feature
   config.include ProjectHelpers
-
   config.include ViewComponent::TestHelpers, type: :component
   config.include Capybara::RSpecMatchers, type: :component